+++ alkos/kernel/src/acpi/osl.tpp
@@ -8,8 +8,7 @@
 #include <pci.hpp>
 
 template <typename T>
-concept ValidUnsigned =
-    std::is_unsigned_v<T> && sizeof(T) <= 4;
+concept ValidUnsigned = std::is_unsigned_v<T> && sizeof(T) <= 4;
 
 template <ValidUnsigned T>
 uacpi_status uacpi_kernel_pci_read(uacpi_handle device, uacpi_size offset, T *value)
